Added check to make sure Path is valid before starting process.

+++ SuperCygwin/Forms/ProcessContainer.cs
@@ -26,26 +26,31 @@ namespace SuperCygwin
         static void em_NewProcess(object sender, EventManager.NewProcessEventArgs e)
         {
             //MessageBox.Show(e.Process.Arguments);
-            if(e.Process.Arguments.StartsWith("/usr/bin/"))
-                if (!File.Exists(e.Process.Arguments.Split(' ')[0].Replace("/usr/bin/", @"C:\cygwin\bin\")) &&
-                    !File.Exists(e.Process.Arguments.Split(' ')[0].Replace("/usr/bin/", @"C:\cygwin\bin\") + ".exe")) 
-                {
-                    if (!File.Exists("cygwin_setup.exe"))
+            if(File.Exists(e.Process.FileName)
+            {
+                if(e.Process.Arguments.StartsWith("/usr/bin/"))
+                    if (!File.Exists(e.Process.Arguments.Split(' ')[0].Replace("/usr/bin/", @"C:\cygwin\bin\")) &&
+                        !File.Exists(e.Process.Arguments.Split(' ')[0].Replace("/usr/bin/", @"C:\cygwin\bin\") + ".exe")) 
                     {
-                        WebClient wc = new WebClient();
-                        try{
-                            wc.DownloadFile("http://www.cygwin.com/setup.exe", "cygwin_setup.exe");
-                        }catch(Exception ex){
-                            MessageBox.Show("Could not download cygwin setup. Please connect to the internet or install " + e.Process.Arguments.Split(' ')[0].Replace("/usr/bin/", ""));
-                            return;
+                        if (!File.Exists("cygwin_setup.exe"))
+                        {
+                            WebClient wc = new WebClient();
+                            try{
+                                wc.DownloadFile("http://www.cygwin.com/setup.exe", "cygwin_setup.exe");
+                            }catch(Exception ex){
+                                MessageBox.Show("Could not download cygwin setup. Please connect to the internet or install " + e.Process.Arguments.Split(' ')[0].Replace("/usr/bin/", ""));
+                                return;
+                            }
                         }
+                        MessageBox.Show("Please install " + e.Process.Arguments.Split(' ')[0].Replace("/usr/bin/", "") + " before attempting to open this preset.");
+                        ProcessStartInfo psi = new ProcessStartInfo("cygwin_setup.exe");
+                        Create(dp,psi);
+                        return;
                     }
-                    MessageBox.Show("Please install " + e.Process.Arguments.Split(' ')[0].Replace("/usr/bin/", "") + " before attempting to open this preset.");
-                    ProcessStartInfo psi = new ProcessStartInfo("cygwin_setup.exe");
-                    Create(dp,psi);
-                    return;
-                }
-            Create(dp, e.Process);
+                Create(dp, e.Process);
+            }else{
+                MessageBox.Show("Error: File Not Found");
+            }
         }
 
         public static ProcessContainer Create(DockPanel dp, string Path, string args = "")
